requirements.txt
matplotlib==3.8.2
numpy==1.26.2
opencv_python==4.8.1.78
Pillow==10.0.1
Pillow==10.1.0
scikit_learn==1.3.2
streamlit==1.29.0
